Updated to correspond with removing hystrix
