+++ metadata/src/main/java/com/redhat/lightblue/rest/metadata/AbstractMetadataResource.java
@@ -51,11 +51,11 @@ import com.redhat.lightblue.rest.metadata.cmd.GetEntityMetadataCommand;
 import com.redhat.lightblue.rest.metadata.cmd.GetEntityNamesCommand;
 import com.redhat.lightblue.rest.metadata.cmd.GetEntityRolesCommand;
 import com.redhat.lightblue.rest.metadata.cmd.GetEntityVersionsCommand;
+import com.redhat.lightblue.rest.metadata.cmd.ReIndexCommand;
 import com.redhat.lightblue.rest.metadata.cmd.RemoveEntityCommand;
 import com.redhat.lightblue.rest.metadata.cmd.SetDefaultVersionCommand;
 import com.redhat.lightblue.rest.metadata.cmd.UpdateEntityInfoCommand;
 import com.redhat.lightblue.rest.metadata.cmd.UpdateEntitySchemaStatusCommand;
-import com.redhat.lightblue.rest.metadata.hystrix.ReIndexCommand;
 import com.redhat.lightblue.rest.util.QueryTemplateUtils;
 import com.redhat.lightblue.util.Error;
 import com.redhat.lightblue.util.JsonUtils;
